+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+* BREAKING: `define` returns a subclass of `PluckMap::Presenter` instead of an instance (@boblail)
+
 ## v1.0.0 (2019 Jul 17)
 
 * FIX: Respect default_scopes for relationships (@boblail)